+++ src/email/message/command/move_.rs
@@ -4,6 +4,8 @@ use clap::Parser;
 use email::message::move_::imap::MoveImapMessages;
 #[cfg(feature = "maildir")]
 use email::message::move_::maildir::MoveMaildirMessages;
+#[cfg(feature = "notmuch")]
+use email::message::move_::notmuch::MoveNotmuchMessages;
 use log::info;
 
 #[cfg(feature = "account-sync")]
@@ -77,6 +79,12 @@ impl MessageMoveCommand {
                         ctx.maildir_for_sync
                             .as_ref()
                             .map(MoveMaildirMessages::new_boxed)
+                    });
+                }
+                #[cfg(feature = "notmuch")]
+                Some(BackendKind::Notmuch) => {
+                    builder.set_move_messages(|ctx| {
+                        ctx.notmuch.as_ref().map(MoveNotmuchMessages::new_boxed)
                     });
                 }
                 _ => (),